feat: restructure Link component
