+++ components/Link/index.ts
@@ -3,9 +3,9 @@
  * Derived from Element Plus v2.9.5
  * https://github.com/element-plus/element-plus/tree/2.9.5/packages/components/link
  */
-import Link from './Link.vue'
+import Link from './src/Link.vue'
 
 export { Link as GLink }
-export * from './link'
+export * from './src/link'
